refactor: submit 네비게이션 로직을 커스텀 훅으로 추상화
